Added expiring_url method to Fog Storage
IMPORTANT: There is a bug in the current release of Fog (1.3.1) that will
return a broken expiring URL when the bucket name is valid for a subdomain.
This is fixed per https://github.com/fog/fog/pull/857 but not released
in the gem yet. I have added currently FAILING tests for these cases that
should pass when a new Fog is released.