Fix a potential bug when user turn on AWS logging
This should fix an error when the user has turn on AWS logging while
and upload a file with non-English filename, as `aws-sdk` logger is not
encoding aware.

The problem was that the string returned from another logger methods
is in `US-ASCII`, while the string returned from Paperclip would be in
`UTF-8`. When the logger calls `#join` on those strings, Ruby would
raise an exception.

This solution is not the best, as it involves a monkey-patching, but
should make sure that the upload goes through without any hiccup. For
the long run, I'm going to send a patch to `aws-sdk` and also make
Paperclip require at least `aws-sdk > 1.4.1`