Create new rake task for changing file options.
Create the rake task paperclip:change_options. This task is invoked the same way
as the refresh tasks (rake CLASS=... paperclip:change_options). It reads a file
config/new_paperclip_options.yml as an ERB YAML file and saves new attachments
for CLASS, including all styles, using the merge of the new options onto the
original options. Existing attachments and models are unaffected.

The purpose of this task is to move existing files to the location/format/styles
needed for a planned change to attachments. You can copy existing files to a new
location using this rake task, then deploy a release which sets the new options
in the model's has_attached_file, and the copied files will be found by
paperclip.

Using the task also simplifies changing hash_secrets.