Add support for generators.
This is a processing step that occurs between processing the original, and
processing the remaining styles. It takes the (optionally processed) original
as input, and outputs any or all of the styles (which are then optionally
processed as normal).

This allows a significant optimization in the common scenario of generating
multiple thumbnails from the source image. The usual processing pipeline could
be replaced with a generator which generates all the output files in a single
invocation, possibly with intermediate downsampling for further performance
gains.