Document how to cleanup after test runs
Adding an example for using Capybara/FactoryGirl to test attachments. I
ran into an issue on my project where the files I was uploading in my
test were getting saved multiple times, leading to hundreds of extra
files before I caught it. I did a bunch of research and found this to be
the most elegant solution (saving files in the testing environment to a
custom location and then deleting that directory after the suite runs).
I thought some clarification in the README would help people deal with
the more easily in the future.