Delegate binmode and binmode? in AbstractAdapter
AbstractAdapter wasn't delegating `binmode` or `binmode?` to the
underlying @tempfile object. `binmode` is useful for cases where the
Adapter object gets handed on to another library expecting an `IO`
object holding image data. We ran into this with Prawn and people using
`prawnto` in conjunction with Paperclip. Prawn wants to ensure the `IO`
is in binmode for embedding the image data.