Fog upgraded how expiring_urls are made
Turns out that between fog 1.22.0 and fog 1.25.0, there was a change to
how fog builds S3 expiring URLs. This changes to test to be the common
point between the two versions. All we need to do here is make sure that
the expiring URL is being constructed.

In the process, this also updates all the appraisal-generated
Gemfiles.

Also, in the process, this meant I had to take away the builder version
restriction. This was put in so that `bundle`ing wouldn't take forever,
but I'm unsire whether this needed to happen with this version or if
builder just needed to be separate from the Gemfile. I'm commimtting
this and asking the people who reported the problem. If you don't see a
revet commit, everything was fine.