Use Updated Version of `aws-sdk` in Docs [ci skip]
The `aws-sdk` development dependency was updated to `~> 1.6` in #1816
to resolve issues with the creation of 0-byte files on S3 (see #1729).
However, there is no runtime dependency on `aws-sdk`, and the
documentation still suggests that users install an earlier version of
`aws-sdk` that produces the problems in the referenced issue.

This commit updates the docs to recommend the use of `aws-sdk` `~> 1.6`
with Paperclip.