Add fog_options to configuration to be passed to fog #create
We found that uploading large files to S3 would result in a socket error
("connection reset by peer") occasionally and lately much more
consistently. In researching this I saw that many people got this error
when uploading too large of a file without multipart chunking. I would
have assumed fog did this automatically but the default chunk size may
be too high. In order to address this I wanted to drop the chunk size to
100MB.

Rather than hard-code this I opted to expose a `fog_option`
configuration option that lets me pass any additional options I want to
the fog's `#create` call. This is similar to the `fog_attributes` option
implemented in CarrierWave which [addresses the same
problem](http://stackoverflow.com/a/11867978/201911).

We've been running this now for a week in production and it seems to
resolve the issue.

https://github.com/thoughtbot/paperclip/pull/2135