Compare to ActiveRecord::VERSION instead of shelling out
`bundle exec rails -v` was taking seconds every time, making the test
suite run very slow. This changes the code to compare the version
against `ActiveRecord::VERSION`, which is going to be the same one as
`rails -v`, making it much, much faster.

Simplifies condition: we no longer support Rails < 4.2, so we test for
`>= "5.0"` or fallback to `4.2` branch.