Correct failing S3 storage spec
This is part of the effort started at https://github.com/thoughtbot/paperclip/issues/2199)
to ensure the test suite is green locally.

This spec was failing on obtaining a S3 instance with an `nil` region
key. Since `Aws::Resource` does a regex match against the region key
passed in, we would fail on `NoMethodError`.

This corrects the issue by setting a region key to an empty string in
the test setup.