Correct attachment processing spec
This is part of the effort started at
https://github.com/thoughtbot/paperclip/issues/2199 to ensure the test
suite is green locally.

The attachment processing spec was failing since the setup required done
in a before block, namely a call to `rebuild_class`, was running
restricted to one of two contexts in the test file.
This was leaving the `Dummy` class in a bad state and affected other tests.

This PR moves the `before` block to run for the both of the contexts.