Add Code Climate Config (#2275)
Adds a Code Climate configuration enabling Fixme, Rubocop, and
Duplication.

The Code Climate RuboCop engine uses a newer version of RuboCop than the
current Hound config, so .hound.yml was modified to work for both. The
only notable change was removing PreferredHashMethods because there is
no syntax for that cop that works with bout Hound and Code Climate.