Cancel all processing if instance contains any errors
Prior to this PR the check for validity happened inside the
run_paperclip_callbacks block, so callbacks where always run.

From the README

NOTE: Post-processing will not even start if the attachment is not
valid according to the validations. Your callbacks and processors will
only be called with valid attachments.

This PR brings the codebase inline with the documented behaviour.

Fixes #2462