Link to Colleen Schnettler's talk
At RailsConf 2019, Colleen Schnettler gave a fantastic talk on her
real-world experience migrating from Paperclip to ActiveStorage, titled
"How to migrate to Active Storage without losing your mind".

In it she goes through the steps before and after a migration, and also
talks through the details of the migration guide, along with some
corrections.

Link to the ConfFreaks video of this talk on YouTube so more
video-minded folks can learn about the migration in a manner that is
more comfortable.