Link to kt-paperclip
We were contacted in early December by Surendra of Kreeti, who explained
that they had forked Paperclip and were maintaining it. We let the
December US holidays go by and then checked in; they're still
maintaining it.

We do not know any of the Kreeti devs and cannot vouch for them as a
team. We did not know of a way to hand over the Rubygem ownership while
expressing that people should remember to audit when upgrading.
Therefore, we felt it best that they have a full fork: new Rubygem gem,
separate repo.

Therefore it makes sense for us to link to Kreeti's fork as an option
for those who cannot move off of Paperclip.

We've linked to `http://kreeti.com` instead of `https` because, as of
this writing, they have left their site open to MITM attacks. Warn the
reader of this so they don't follow a cleartext link without necessary
preparation.